added type casting in intent extras in TaskActivity and TaskAdapter classes. fixed the extras sent int TaskActivity

+++ app/src/main/java/com/example/android/employeesmanagementsoftware/taskDB/TasksAdapter.java
@@ -62,7 +62,7 @@ public class TasksAdapter extends RecyclerView.Adapter<TasksAdapter.TasksViewHol
             {
                 id = tasks.getId();
                 Intent in = new Intent(context, TaskActivity.class);
-                in.putExtra("task_id",id);
+                in.putExtra("task_id",(long)id);
                 context.startActivity(in);
             }
         });
